chore: Auth와 User 기능 분류
