+++ src/main/java/org/example/unpluguserservice/controller/AuthController.java
@@ -6,7 +6,6 @@ import org.example.unpluguserservice.common.ApiResponse;
 import org.example.unpluguserservice.dto.UserRequestDto;
 import org.example.unpluguserservice.dto.UserResponseDto;
 import org.example.unpluguserservice.service.AuthService;
-import org.example.unpluguserservice.service.UserService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,24 +14,23 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AuthController {
     private final AuthService authService;
-    private final UserService userService;
     private final PasswordEncoder passwordEncoder;
 
     @PostMapping("/signup")
     public ApiResponse<UserResponseDto> registerUser(@Valid @RequestBody UserRequestDto request){
-        UserResponseDto response = userService.createUser(request, passwordEncoder);
+        UserResponseDto response = authService.createUser(request, passwordEncoder);
         return new ApiResponse<>(true, "회원가입이 정상적으로 완료되었습니다.", response);
     }
 
     @GetMapping("/check/username")
     public ApiResponse<String> checkUsername(@RequestParam("username") String username){
-        userService.checkUsableUsername(username);
+        authService.checkUsableUsername(username);
         return new ApiResponse<>(true, "사용 가능한 아이디입니다.", username);
     }
 
     @GetMapping("/check/nickname")
     public ApiResponse<String> checkNickname(@RequestParam("nickname") String nickname){
-        userService.checkUsableNickname(nickname);
+        authService.checkUsableNickname(nickname);
         return new ApiResponse<>(true, "사용 가능한 닉네임입니다.", nickname);
     }
