+++ src/main/java/org/example/unpluguserservice/service/AuthService.java
@@ -1,6 +1,8 @@
 package org.example.unpluguserservice.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.unpluguserservice.dto.UserRequestDto;
+import org.example.unpluguserservice.dto.UserResponseDto;
 import org.example.unpluguserservice.entity.RefreshToken;
 import org.example.unpluguserservice.entity.User;
 import org.example.unpluguserservice.repository.RefreshTokenRepository;
@@ -9,7 +11,9 @@ import org.example.unpluguserservice.util.JwtUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 
 @Service
@@ -22,6 +26,33 @@ public class AuthService {
 
     @Value("${jwt.refresh-token-expiration-ms}")
     private long refreshExpMs;
+
+    @Transactional
+    public UserResponseDto createUser(UserRequestDto request, PasswordEncoder passwordEncoder){
+        User user = User.builder()
+                .username(request.getUsername())
+                .name(request.getName())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .nickname(request.getNickname())
+                .createdDate(LocalDateTime.now())
+                .build();
+        User savedUser = userRepository.save(user);
+        return UserResponseDto.fromEntity(savedUser);
+    }
+
+    @Transactional(readOnly = true)
+    public void checkUsableUsername(String username){
+        if(userRepository.findByUsername(username).isPresent()){
+            throw new RuntimeException("이미 존재하는 아이디입니다.");
+        }
+    }
+
+    @Transactional(readOnly = true)
+    public void checkUsableNickname(String nickname){
+        if (userRepository.findByNickname(nickname).isPresent()){
+            throw new RuntimeException("이미 존재하는 닉네임입니다.");
+        }
+    }
 
     public LoginResponse login(String username, String password){
         User user = userRepository.findByUsername(username)
